requirements.txt
python-pptx
PyPDF2
fitz
pytesseract
from pdf2image import convert_from_path
tempfile